New feature with threads

+++ types.go
@@ -364,8 +364,8 @@ type Message struct {
 	// From is a sender, empty for messages sent to channels;
 	//
 	// optional
-	MessageThreadID int `json:"message_thread_id,omitempty"`
-	From *User `json:"from,omitempty"`
+	MessageThreadID int64 `json:"message_thread_id,omitempty"`
+	From            *User `json:"from,omitempty"`
 	// SenderChat is the sender of the message, sent on behalf of a chat. The
 	// channel itself for channel messages. The supergroup itself for messages
 	// from anonymous group administrators. The linked channel for messages
